requirements.txt
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.0
google-api-python-client>=2.130.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
sqlalchemy>=2.0.0
APScheduler>=3.10.0
streamlit>=1.35.0
plotly>=5.22.0
reportlab>=4.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.32.0
beautifulsoup4>=4.12.0
